+++ src/main/factories/presentation/placeListFactory.tsx
@@ -4,12 +4,12 @@ import { Actions, Routes, navigator } from '~/main/navigation';
 import { ListPlaces, ListRecommendations } from '~/domain/useCases';
 import { PlaceModel, RecommendationModel } from '~/domain/models';
 import { Place } from '../../../presentation/components/cardList';
-import { RecommendationProps } from '../../../presentation/recommendation/components/listRecommendation';
 import { StackParams } from '../../navigation/navigation';
 import PlaceList from '../../../../src/presentation/placeList';
 import usePlaceList, {
   Origin,
 } from '../../../../src/presentation/placeList/usePlaceList';
+import getPlacesByOrigin from './helpers/getPlacesByOrigin';
 
 class ListRecommendationsDAO implements ListRecommendations {
   async list(): Promise<RecommendationModel[]> {
@@ -55,7 +55,7 @@ const PlaceListFactory = ({}: Props) => {
   const [places, setPlaces] = useState<Place[]>([]);
 
   const getPlaces = async () => {
-    const response = await getListByOrigin(
+    const response = await getPlacesByOrigin(
       Origin.Places,
       new ListRecommendationsDAO(),
       new ListPlacesDAO(),
@@ -74,43 +74,6 @@ const PlaceListFactory = ({}: Props) => {
     places,
   });
   return <PlaceList {...viewModel} />;
-};
-
-export class RecommendationsMapper {
-  constructor(private readonly recommendations: RecommendationProps[]) {}
-
-  toPlaces = (): Place[] => {
-    return this.recommendations.map((recommendation) => {
-      return {
-        amountOfReviews: '',
-        imageUrl: recommendation.imageUrl,
-        location: recommendation.location,
-        myDistanceOfLocal: recommendation.myDistanceOfLocal,
-        rating: recommendation.rating,
-        title: recommendation.title,
-      };
-    }) as Place[];
-  };
-}
-
-export const getListByOrigin = async (
-  by: Origin,
-  listRecommendations: ListRecommendations,
-  listPlaces: ListPlaces,
-  location: { lat: string; long: string },
-): Promise<Place[]> => {
-  if (by === Origin.Recommendations) {
-    const response = await listRecommendations.list();
-    return new RecommendationsMapper(response).toPlaces();
-  }
-
-  if (by === Origin.Places) {
-    if (location) {
-      return await listPlaces.list(location);
-    }
-  }
-
-  return [];
 };
 
 export default PlaceListFactory;
